fix float to int

+++ train.py
@@ -27,7 +27,7 @@ def main(config=None):
         
         logger.info(f"Constructing model with configs: {config}")
         # divide embed_dims by 2 to get position_embed_dims
-        model = ScoreModel(embed_dims=config.embed_dims, num_conv_layers=config.num_conv_layers, position_embed_dims=config.embed_dims/2, tmin=0.001, tmax=1000000.0)
+        model = ScoreModel(embed_dims=int(config.embed_dims), num_conv_layers=int(config.num_conv_layers), position_embed_dims=int(config.embed_dims/2), tmin=0.001, tmax=1000000.0)
         
         total_params = sum([p.numel() for p in model.parameters()])
         logger.info(f"Model has {total_params} params")
